refactor: fix metadata resetting after submit in Panel, refactor get object prefetch into custom hook
