+++ src/components/inputs/skylarkObjectFieldInput/skylarkObjectFieldInput.component.tsx
@@ -281,8 +281,6 @@ export const SkylarkObjectFieldInput = (
     },
   };
 
-  console.log(field, inputProps);
-
   return (
     <div className="group/input-field mb-4 text-sm">
       <SkylarkObjectFieldInputLabel
